refactor(checkout): extract promotional code logic to a custom hook
