+++ sonar-project.properties
@@ -3,7 +3,7 @@ sonar.projectKey=internxt_drive-web
 sonar.projectName=drive-web
 sonar.sources=src
 sonar.tests=src,test/unit
-sonar.coverage.exclusions=src/**/*.tsx,src/config.ts, **/*.errors.ts, **/plugins/*.ts
+sonar.coverage.exclusions=src/**/*.tsx,src/config.ts, **/*.errors.ts, **/plugins/*.ts, **/constants.ts
 sonar.test.inclusions=src/**/*.test.ts,src/**/*.test.tsx,test/unit/**/*.test.ts
 sonar.javascript.lcov.reportPaths=coverage/lcov.info
 sonar.exclusions=**/*.json,**/assets/**,**/*.scss,**/src/setupTests.ts
